remove last step of product form
